multi-slot refactoring

+++ assets/js/data/localStorage.js
@@ -1,15 +1,19 @@
 /* exported getBuildingPC, setBuildingPC, deleteBuildingPC */
-
+const piecesMulti = ['ram', 'rom'];
 const getBuildingPC = () => {
    return JSON.parse(localStorage.getItem('buildingPC'));
 };
 
 const setBuildingPC = (piece, idDiv) => {
    let myPC = getBuildingPC();
+
    if (!myPC) myPC = generateBody();
    if (piece) {
-      if (piece.type === 'ram')myPC = multipleSlots(piece, idDiv, myPC);
-      else myPC[piece.type] = piece;
+      if (piecesMulti.includes(piece.type)) {
+         myPC = multipleSlots(piece, idDiv, myPC);
+      } else {
+         myPC[piece.type] = piece;
+      }
    }
    localStorage.setItem('buildingPC', JSON.stringify(myPC));
 };
@@ -28,34 +32,30 @@ const generateBody = () => {
    };
 };
 
-const multipleSlots = (piece, idDiv, myPC) =>{
-   if (piece.type === 'ram') {
-      if (!myPC.ram) myPC.ram = [];
-      myPC.ram.push({...piece, div: idDiv});
-   }
+const multipleSlots = (piece, idDiv, myPC) => {
+   if (!myPC[piece.type]) myPC[piece.type] = [];
+   myPC[piece.type].push({...piece, div: idDiv});
+
    return myPC;
 };
 
 const deleteBuildingPC = (deletePiece) => {
    // eslint-disable-next-line prefer-const
    let myPC = getBuildingPC();
-
-   switch (deletePiece.type) {
-   case 'ram':
+   if (piecesMulti.includes(deletePiece.type)) {
       myPC[deletePiece.type] = deleteArray(deletePiece, myPC);
-      break;
-   default:
+   } else {
       myPC[deletePiece.type] = null;
    }
 
    localStorage.setItem('buildingPC', JSON.stringify(myPC));
 };
 
-const deleteArray = (piece, myPC) =>{
-   const anws = myPC[piece.type].filter((element) =>{
+const deleteArray = (piece, myPC) => {
+   const answer = myPC[piece.type].filter((element) => {
       piece.div = element.div;
       return JSON.stringify(element) !== JSON.stringify(piece);
    });
-   if (anws.length === 0) return null;
-   else return anws;
+   if (answer.length === 0) return null;
+   else return answer;
 };
